Extracted deforestation trajectory logic for reusing
